COMMERCE-6422 move to util module
