+++ modules/apps/commerce/commerce-test/src/testIntegration/java/com/liferay/commerce/order/engine/test/CommerceOrderEngineTest.java
@@ -48,6 +48,9 @@ import com.liferay.commerce.service.CommerceShipmentItemLocalService;
 import com.liferay.commerce.service.CommerceShipmentLocalService;
 import com.liferay.commerce.test.util.CommerceTestUtil;
 import com.liferay.commerce.test.util.context.TestCommerceContext;
+import com.liferay.commerce.test.util.order.status.Test1CommerceOrderStatusImpl;
+import com.liferay.commerce.test.util.order.status.Test2CommerceOrderStatusImpl;
+import com.liferay.commerce.test.util.order.status.Test3CommerceOrderStatusImpl;
 import com.liferay.petra.lang.CentralizedThreadLocal;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Company;
